Fix infinite redirect bug
If the OPENID_INJECT_GET_PARAM was set, and the OPENID_SSO_SERVER_URL
contains a formatting string, the server url will be invalid and can
result in infinite redirects. Removing the requirement that there be
GET parameters at all solves the problem (openid_url gets properly set to
None, so that the normal login flow begins).